+++ GoRogue/GoRogue.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <!-- Basic package info -->
-    <TargetFrameworks>netstandard2.1;netcoreapp3.1;net5.0;net6.0;net7.0;net8.0</TargetFrameworks>
+    <TargetFrameworks>netstandard2.1;netcoreapp3.1;net5.0;net6.0;net7.0</TargetFrameworks>
 	<RootNamespace>GoRogue</RootNamespace>
 	<Authors>Chris3606</Authors>
 	<Copyright>Copyright © 2023 Christopher Ridley (Chris3606)</Copyright>
